vFire: Improved Integration (#1549)
With TTT2 v0.13.0 we started supporting vFire in TTT2. This pullrequest
improves the integration by reworking the game effects function to use
the vFire particles when vFire is installed. I added a `0.5` modifier to
the vFire fire particles as they were quite large and almost
overwhelming. Now they are more tuned to TTT2. However I doubled the
damage ticks so it more closely matches that of normal TTT.

This PR also adds a hook replacement to the vFire explosion handling
reported in #1478. The issue here is that it was applied to all
entities, but it shouldn't be applied to players. I couldn't really test
this one as I was unable to reproduce the issue due to loading order.
See the issue for more details.

Also maybe it is a good idea if @EntranceJew could take a look at this
as well because he worked more with vFire than I did.

vFire docs:
https://docs.google.com/document/d/1GbSrWMqJYI0f4pP0pYBiG2TsKUCXifQLjCqC0qHA9a0/edit