Button Fixes and TargetID Integration (#1580)
Fixes #1557

Adds TargetID to leavers ...

![image](https://github.com/user-attachments/assets/0d73ac31-d1ea-4d30-9b52-ec346669317b)

... and buttons

![image](https://github.com/user-attachments/assets/15aa86fa-9850-4f9a-a68b-47d5da47ca0d)

Note: The fix of this issue was done by making the lever and buttons
non-solid. This has to be done both on server and client. After that was
done, the addition of targetID consists only of 20 lines of code.

Note 2: The button and lever entities are non-detectable on the client
as they are only a base entity there. Therefore it has to be synced from
the server.

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>