Player/Roles: Added Fallback Colors (#1597)
Fixed an issue @sbzlzh noticed where there are errors on color access in
the first round. Strangely this only happens if it is the first round
and ROLE_NONE. This fix looks kinda hacky imho, but the issue here seems
to be rooted way deeper. I think it has probebaly something to do with
our role syncing and that the default is alredy ROLE_NONE, so a role
change is never triggered.

Moreover the same hacky solution was already implemented for the default
role color call. In my opinion this solution is fine, but we should keep
this issue in mind in case we refactor the role syncing at some point.

This issue also probably existed since our introduction of ROLE_NONE.
Nobody noticed it, because the default player amount is 2. And with two
players, there are roles assigned.