PlayerReady: Extended Syncing (#1602)
As discussed with @mexikoedi, I made `GM:TTT2PlayerReady` fully shared
to be used in his #1600 PR. This might cause issues with third party
addons, but at least for our addons, the change is fine. And I'm pretty
sure nobody else is using that hook anyway.

To test:
- is `GM:SetupMove` only called for the local player on clients? I think
so, but it has to be confirmed
- does this work at all?
- is every player on the client marked as ready, no matter the order in
which the players connected?