Buttons: Added Serverside Fallback (#1608)
Since there were still issues with clientside use detection (at least on
the map `ttt_mc_jondome` one button refused to work), I added a
serverside fallback. If the client doesn't detect any entity on use
press, the server uses the engine function to detect any use entity
close by. This works.

Only downside is that it triggers uses for entities where the targetID
is just slightly out of range. However this is how it has been in 0.13.2
and earlier as well.

Update: In the last commit I implmented `IsUsableEntity` from the
sourceSDK in lua. All important links to the original code are in the
added code. I had to do this because enties without a `+use` assigned
could block the trace. Now it uses the aforementioned serverside
fallback if an entity has no use.


![image](https://github.com/user-attachments/assets/e5943562-5cee-44d8-ae11-9d1e3e583d77)


https://discord.com/channels/442107660955942932/467653559408656384/1285124499401998356