Lua ent use fix (#1641)
This fixes #1634 and #1635

The issue this PR fixes is that previously the fallback was called for
all SENTs (scripted entities) because they lack any engine related use
flags. Therefore the clientside trace was completely ignored for this
kind of entity.

I added support for our SENTs to `entmeta:IsUsableEntity`. This might
look a bit hacky, but let me explain:
- First, the obvious, if an entity has `.CanUseKey` set to true, it has
a use functionality attached. This is a good check, because this is a
flag that was needed in vanilla TTT as well, so it has good
compatibility.
- Second, there are special cases for weapons and ragdolls. This is a
bit hacky indeed. However it is handled the same way in the server use
handler, so I guess it is fine


![image](https://github.com/user-attachments/assets/2e94940b-c365-4d17-9a7f-986e8741aee9)

---------

Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>