Loadingscreen: Fixed Disabling (#1643)
Fixes #1622

Fixes disabling of the loadingscreen breaking the game. Also I fixed the
loading screen displaying one fewer rounds than are actually left
because the game counts to 0 and it makes no real sense that in the last
round are 0 rounds left.

I also made sure that the prepare time is set to the correct value if
the loading screen is disabled.