fix(map): Unconditionally create gm, ttt and ttt2 prefixes
We can reasonably assume `gm` maps to be present on the server the same
goes for `ttt` and `ttt2` maps.
Even if either or both of them are not currently installed on the server
creating these unconditionally would allow an admin to configure their
behaviour in that case.

Creating these convars unconditionally instead of dynamically has the
added upside that if an admin would change the value to non-default the
value value will be set on file read and won't trigger `FCVAR_NOTIFY`.
(An "active" convar change mid-round will still trigger `FCVAR_NOTIFY`
as normal)

Creating convars dynamically if we find other map prefixes is still done
and is still affected by the above caveat.