feat(none): Better deprecation message
This introduces a designated function to inform about deprecations.

With this we now have a single interface for deprecations with the
following key points:
* Only players with `developer 1` or higher will see an error
* Simple function interface; 2 strings explaining the deprecation
* Automatically point to file where the deprecated function was used

The main motivation behind this change was to make our lives easier when
deprecating functions and when troubleshooting players' issues.

In the former case we now don't need to think about how to inform about
a deprecation, we simply use this new function.
In the latter case we don't need to tell players to ignore these
deprecation warnings because they won't see them. Also problem reports
will contain less errors overall.