TargetID: Make `ttt_idenfity_body_woconfirm` replicated; move and clean up additional ConVars surrounding corpses/ragdolls (#1692)
In #1684, `targetid.HUDDrawTargetIDRagdolls` was updated to check if the
`ttt_idenfity_body_woconfirm` ConVar was enabled to determine what hint
gets displayed.

Problem is, this ConVar only exists on the server, and the TargetID
library is entirely clientside. This resulted in Lua errors being thrown
on every frame the function was called and managed to reach that check.

This PR adds an additional table to the `CORPSE` namespace for the
purpose of storing ConVar objects related to corpses and bodysearching.
The following `CreateConVar` calls have been moved from `sv_main`:
- `ttt_idenfity_body_woconfirm` (to `sh_corpse`, with `FCVAR_REPLICATED`
added)
- `ttt2_confirm_team` (to `sv_corpse`)
- `ttt2_confirm_killlist` (to `sv_corpse`)

I've also moved the `CreateConVar` calls for `ttt_announce_body_found`
and `ttt_ragdoll_collide` into the new table for consistency, and so
scripts for other parts of the gamemode can access the ConVar objects
directly without having to store their own separate copies (or calling
`GetConVar` repeatedly). I've also already adjusted all existing checks
for these ConVars across the gamemode accordingly.

The changelog entry is worded the way it is because I just can't think
of any other way to describe the results of the changes without going
into a level of detail that'd be unnecessary for an end-user, but still
useful information for addon developers.