Markervision fix: Fixed registry table being able to have duplicates (#1712)
I discovered that when new marker vision objects are made with
`markerVision.Add`, it doesn't clear out existing versions of the same
object (matching ent and indentifier). This leads to the
`markerVision.registry` table being filled up with objects that do not
clear themselves, and makes `markerVision.Get` potentially return
obsolete markers (mainly observed serverside).

This PR resolves this possible bloat issue in the registry. This also
fixes an issue where marker vision objects can incorrectly change to
being visible to team or visible to role (clientside), when it's
supposed to be visible to everyone.