Gameloop: Split round and time (#1710)
This PR will tackle https://github.com/TTT-2/TTT2/issues/1707

Not finished yet, still have to do all LANG files and some testing +
cleanup if neccessarry.

Would love some input on this though!

Only "real" problem I have so far is the settings in the F1 menu: I can
not "disable" the round time or the round count setting because the
"master" is never null but always a value.. Should not really be
confusing though.