Continous Use: Reimplemented (#1645)
Fixes #1636

Due to the use rework, continuous use was broken. This PR fixes it by
adding a timer that is called as long as the use key is pressed and the
entity is close by.
Most of these changes are just code being moved into another function,
the timer part is new though

---------

Co-authored-by: ZenBre4ker <72046466+ZenBre4ker@users.noreply.github.com>