vgui: Reworked inheritance (#1722)
This PR adds no new code or new functionality, it only moves code around
in files so that everything that inherits from buttons can use (server)
convars. Therefore the code from the ttt2 checkboxlabel was moved to
button. This caused this change to track the underlying button in those
labels:


![image](https://github.com/user-attachments/assets/afeef60e-1d7d-4b66-b44f-15f4ec827ce8)

Moreover I added our tooltip code from the panel to the label as well,
as both of these are a starting point for the inheritance that can't be
merged.

I'd appreaciate a quick merge so that #1715 can be finished.

Also note, that there are MANY things that could be further improved by
improving the inheritance. This only tackles a small part of that.