HUD: Added Team name next to Role name (#1726)
Hello!

I have encountered a problem on my friends server where if we have many
roles installed, some players who join will not know what every single
role is, and they will become confused, especially with what team they
are supposed to play on.

This new change will allow players to see what team they are on,
immediately next to their role name. I know there is a team ICON next to
the time remaining in the match on the top of the screen, but our
players either never even noticed it, or did not know what it meant.

I have included a toggle in the F1 menu, under HUD Switcher so that
players who feel like they don't need to see the team can turn it off
and enjoy the game their way.

If you guys have any changes to be made I would love to hear it, I think
this change will help new player retention!

Here are some previews of how it will look, including the F1 menu and an
example with a long role name!

![f1_menu](https://github.com/user-attachments/assets/a67e867d-53e3-4d57-b2b4-d5af7b15060b)

![short_name](https://github.com/user-attachments/assets/12be0364-4b44-465c-83ab-aaf3855616ac)

![long_name](https://github.com/user-attachments/assets/fdd3e46e-edfd-4274-96aa-d2ac82da177c)

---------

Co-authored-by: Tim Goll <github@timgoll.de>