Added Win Score Event (#1733)
Added win score to finish event with parameter for amount of score in
role base and LANG entries

I felt like it was weird that no score was gained for winning,
especially for roles that cant gain score as quickly in other ways. This
adds an event for that where score amount can also be configured for
every role.