vGui: Fixed Clientside ConVars & Small Misc Fixes (#1742)
fixes #1735

This issue was created since the cvar is now tracked in the "main
button" not the checkbox button anymore. This worked fine for serverside
settings with custom handling, as well as client settings that use
function wrappers to set the convar (like vskin for example).

This change now correctly sets the stored convar. It feels a bit hacky
to me to be honest, but only a full rewrite would resolve this. So it is
probably fine.

I also noticed that [this requested
change](https://github.com/TTT-2/TTT2/pull/1724#discussion_r1921642625)
was not applied to all function calls, so I renamed the one remaining.

Last but not least, I noticed that the border radius was substracted
twice.