Role distribution inspection (#1714)
This PR adds an admin UI to inspect how and why role distribution
selects certain roles, and why it distributes those roles to players.

The goal here is to assist admins in understanding the role distribution
process, so they can more effectively tweak the parameters involved, and
better see their effects.

This also includes a graphic showing player weights during role
distribution (if derandomization is enabled) which can assist debugging
that as well as help intuition when tweaking it.

Some of the graphics are a touch jank, and in particular I wish there
was a good way to get an icon in the label of a collapsible form, but
just role names probably works fine anyway.

This PR is based on (and soft depends-on) #1702. That dependency can be
removed if necessary.

Screenshots:

![Screenshot_20250111_221601](https://github.com/user-attachments/assets/3238a7ae-68ad-4e32-89f2-3c88ae47faea)

![Screenshot_20250111_222215](https://github.com/user-attachments/assets/623eac41-db13-45d0-a8aa-46815bc72603)

![Screenshot_20250111_222252](https://github.com/user-attachments/assets/cae1e417-d63b-44fe-9bc0-73be97007d0c)

![Screenshot_20250111_222312](https://github.com/user-attachments/assets/1f4ca740-f9c7-4ded-be63-da09b73d98b3)

![Screenshot_20250111_222324](https://github.com/user-attachments/assets/b34f0688-5423-4bf9-8bba-8fd312631056)

![Screenshot_20250111_222328](https://github.com/user-attachments/assets/2e9c9ba4-ddfa-43ea-b492-c5a59f0b8bcc)

![Screenshot_20250111_221616](https://github.com/user-attachments/assets/e63513b4-5852-4a6e-a578-abbbf9b95ddd)

![Screenshot_20250111_221629](https://github.com/user-attachments/assets/980540be-54e9-4268-93cd-687ea3be8611)

![Screenshot_20250111_221656](https://github.com/user-attachments/assets/e1f1798f-4e68-4ad2-b54c-61751d64d714)

![Screenshot_20250111_221731](https://github.com/user-attachments/assets/7884d52a-c323-41ed-8d9b-137627c8574c)

![Screenshot_20250111_221747](https://github.com/user-attachments/assets/90405fac-f055-4d97-93ea-da771c6bbfbc)

---------

Co-authored-by: Tim Goll <github@timgoll.de>
Co-authored-by: Histalek <16392835+Histalek@users.noreply.github.com>