VoiceChat: Added option for team proximity voicechat (#1811)
This adds a new convar/menu option to enable/disable proximity voice
chat for the team voice chat.

This tackles https://github.com/TTT-2/TTT2/issues/1810

I was NOT yet able to test this! For me it seems 100% fine but if anyone
can/wants to test it before merging, please do so.