fix(keys): Refactor to fix continuous use handling
We refactored the "use key handling" to correctly
work with entities that use the continuous use
functionality, like the health station.

Why?

Previous efforts to consolidate "use key handling"
led to a regression where continuous use
did not work as intended anymore when directly
looking at such an entity.

Though, it still worked when using the use key
in the proximity of the entity without directly
looking at it or running into it with the use
key held down, because the server then handled
the continuous use by itself.

This is behavior that the client cannot "block",
so we needed to refactor this code to use the
server-side continuous use handling as its only
source of truth again.

The root cause for this issue was that the client
suppressed further use handling after sending
the net message to the server, which caused the
server-side `KeyPressed` check to return false in
the timer.
Simply changing this would not be sufficient,
as we need still to suppress this handling for the
other use-cases, like weapon pickup or remote use.

Changes:

We mainly refactored the server-side
"use key handling" and client-side key suppression
to fix this issue and simplify the code.
Also, the continous use execution frequency is now
every tick again, instead of every 0.1 seconds.

The actual continuous execution is now handled by
solely by the server again.

The `TTT2PlayerUseEntity` net message code was
refactored to only handle its special cases like
spectator use, remote use and weapon pickup.
We send this message from within the bind handler
on the client if applicable.
This code is still necessary, as these are cases
where the normal use handling is not triggered
by the server due to the items being out of reach
or may not be correctly prioritized e.g.
the wrong weapon being picked up instead of
the intended looked at entity.

The corpse search and old `UseOverride` handling
was moved back to the original TTT way of handling
it, by utilizing the `KeyRelease` hook.
We did not modify the code here and merely copied
over the original code again.

Removed `ENT.CanUseKey` from placeable entities
as it is not needed anymore with the new handling.

Removed now obsolete `IsSpecialUsableEntity` and
`IsUsableEntity` functions.

Refactored `ClientUse` function to be client only
and always block further use handling.
This function is generally useful for client-side
only use key handling, e.g., for UI interactions.
E.g. the C4 uses this to open the C4 menu
when the use key is pressed while looking at it.