Fixed guns not applying themselves as their damage inflictor for FireBullets (#1819)
I've noticed that guns in TTT2 don't set their Inflictor as themselves
for FireBullets ([which is a new-ish
addition](https://wiki.facepunch.com/gmod/Structures/Bullet#Inflictor)),
so I've fixed that.

I'd like to note it seems `util.WeaponFromDamage` was partially used as
a workaround for FireBullets not setting an Inflictor, do people think
it's still needed?