TargetId: Let GM:TTTModifyTargetedEntity modify the target distance (#1828)
You can modify the target entity, but the distance value would remain
the same since it couldn't be changed. This can make targetids to appear
and disappear unexpectedly if the target entity was changed.

I've made `GM:TTTModifyTargetedEntity` read a second return value for
distance so devs can correct the distance or do as they please with it.
I've checked and it doesn't let you use entities from far away if you
force target them with distance set to 0.