Markervision: Fixed null entity error (#1835)
This PR fixes #1834 .
The null entity error was introduced in #1825 .

I tested it and I couldn't reproduce the error from the above issue
again.

Error reason:
markerVision.Remove is called twice.
The first time ent is valid but the second time it's a null entity and
that caused the error.